Autocomplete fills in place name from region in table when clicked   TODO: dismiss table when user touches outside table   TODO: save lon and lat into variables for API call instead of names   TODO: modify API to check what form query is in and handle lat/lon requests   TODO: clear previous search results when switching origin/destination   TODO: improve styling

+++ StreetWise/JHDirectionSearchResults.m
@@ -44,7 +44,7 @@
 }
 
 - (MKPolyline *)createPolyLineFromPoints:(NSArray *)pathPoints {
-    int numberOfPoints = pathPoints.count;
+    NSInteger numberOfPoints = pathPoints.count;
     CLLocationCoordinate2D coords[numberOfPoints];
     
     for (int i = 0; i < numberOfPoints; i++) {
